requirements.txt
requests
twitch-python
selenium
python-chess
